Renamed: component PickerListType.vue
