+++ apps/frontend/src/components/page-products/PickerListType.vue
